requirements.txt
asgiref==3.9.1
Django==5.2.4
djangorestframework==3.16.0
sqlparse==0.5.3
gunicorn>=20.1
python-dotenv>=1.0
psycopg2-binary>=2.9
